refactor(folders): remove standard columns, fontsize and counties
