+++ tavla/app/(admin)/edit/[id]/components/DuplicateBoard/actions.ts
@@ -1,11 +1,10 @@
 'use server'
-import { getOrganizationIfUserHasAccess } from 'app/(admin)/actions'
 import { getFormFeedbackForError } from 'app/(admin)/utils'
 import { initializeAdminApp } from 'app/(admin)/utils/firebase'
 import { getUserFromSessionCookie } from 'app/(admin)/utils/server'
 import admin, { firestore } from 'firebase-admin'
 import { redirect } from 'next/navigation'
-import { TBoard, TOrganization, TOrganizationID } from 'types/settings'
+import { TBoard, TOrganizationID } from 'types/settings'
 import * as Sentry from '@sentry/nextjs'
 
 initializeAdminApp()
@@ -13,9 +12,6 @@ initializeAdminApp()
 export async function duplicateBoard(board: TBoard, oid?: TOrganizationID) {
     const user = await getUserFromSessionCookie()
     if (!user) return getFormFeedbackForError('auth/operation-not-allowed')
-
-    let organization: TOrganization | undefined
-    if (oid) organization = await getOrganizationIfUserHasAccess(oid)
 
     let createdBoard = null
 
@@ -26,10 +22,7 @@ export async function duplicateBoard(board: TBoard, oid?: TOrganizationID) {
                 ...board,
                 meta: {
                     ...board.meta,
-                    fontSize:
-                        board.meta?.fontSize ??
-                        organization?.defaults?.font ??
-                        'medium',
+                    fontSize: board.meta?.fontSize ?? 'medium',
                     created: Date.now(),
                     dateModified: Date.now(),
                 },
