+++ tavla/app/(admin)/hooks/useCountiesSearch.ts
@@ -1,10 +1,8 @@
 import { NormalizedDropdownItemType } from '@entur/dropdown'
 import { useCallback, useEffect, useState } from 'react'
-import { TOrganization, TOrganizationID } from 'types/settings'
 import { fetchCounties } from 'app/(admin)/utils/fetch'
-import { getOrganizationIfUserHasAccess } from 'app/(admin)/actions'
 
-function useCountiesSearch(oid?: TOrganizationID) {
+function useCountiesSearch() {
     const [countiesList, setCountiesList] = useState<
         NormalizedDropdownItemType[]
     >([])
@@ -15,16 +13,6 @@ function useCountiesSearch(oid?: TOrganizationID) {
     useEffect(() => {
         fetchCounties().then((res) => setCountiesList(res))
     }, [])
-
-    useEffect(() => {
-        getOrganizationIfUserHasAccess(oid).then((o?: TOrganization) => {
-            if (!o) return setSelectedCounties([])
-            const defaultCounties = countiesList.filter((c) =>
-                o.defaults?.counties?.includes(c.value),
-            )
-            setSelectedCounties(defaultCounties)
-        })
-    }, [countiesList, oid])
 
     const counties = useCallback(() => countiesList, [countiesList])
